Failed attempt at setting manual mipmaps
Based on three.js' example at http://threejs.org/examples/webgl_materials_texture_manualmipmap.html
Setting .mipmaps to array of canvases each a smaller power of two in
size. But this fails with 256 errors:

WebGL: drawElements: texture bound to texture unit 0 is not renderable.
It maybe non-power-of-2 and have incompatible texture filtering or is
not 'texture complete'. Or the texture is Float or Half Float type with
linear filtering while OES_float_linear or OES_half_float_linear
extension is not enabled.